Documentation and saving the Maze
Added functions to save the maze as a PDF file in a folder. The program looks into the folder to see how many files are there (assuming all the files there are maze pdfs) and generates a maze filename.  The program then redraws the maze in B&W and saves the frame as a PDF.

Added a header and a bunch of comments to the code.